app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow==2.15.0
pillow==10.1.0
jinja2==3.1.2
python-multipart==0.0.6
